Applied min-price to the /gts sell <type> <price>
Was told there was a bug where the min price was not applied to selling when using a command. Haven't looked into the GTS source before so getting everything didn't work quite well. But this should help get the ball rolling for a good portion of the implementation.

Found some max price values in ConfigKeys but couldn't find anything regarding min-price besides the Minable interface having a method for calculating the min price. This will need to be edited but is most likely more helpful than just an issue =)